chore(db): fase 2 migracion postgres - traduccion V1..V15
Traduce las 15 migraciones Flyway de MySQL a sintaxis PostgreSQL:
- TIMESTAMP(6) DEFAULT CURRENT_TIMESTAMP(6) [ON UPDATE ...] -> timestamptz
  DEFAULT now() (la app fija created_at/updated_at via @PrePersist/@PreUpdate;
  Instant mapea a timestamptz en PostgreSQLDialect).
- TINYINT -> smallint en recipe_ratings (V10) y en RecipeRatingEntity.stars.
- V11: quita "AFTER display_name" (no soportado en Postgres).
- V13: CURRENT_TIMESTAMP(6) -> now(); el resto es ANSI, sin cambios.
- CHAR(36), DECIMAL, INT, BIGINT, BOOLEAN, DATE, indices, PK/FK/UNIQUE/CHECK
  sin cambios (validos en Postgres).

Tests migrados a PostgreSQL real (no hay Docker/Testcontainers en el entorno):
- application-test.yml y DevDataSeederTest apuntan a DB de test via env
  (DB_TEST_URL/USERNAME/PASSWORD), driver org.postgresql, ddl-auto=validate.
- Eliminada la dependencia H2 (ya no se usa).

Validado en sesion: mvn -DskipTests test-compile OK (main + test compilan).
NO validado aun: mvn test contra Postgres real (pendiente DB de test, BLOQUE 4).

+++ backend/src/main/java/org/gipsybuho/recetasfamiliares/ratings/RecipeRatingEntity.java
@@ -37,7 +37,7 @@ public class RecipeRatingEntity {
     @JoinColumn(name = "user_id", nullable = false, columnDefinition = "CHAR(36)")
     private UserEntity user;
 
-    @Column(nullable = false, columnDefinition = "TINYINT")
+    @Column(nullable = false, columnDefinition = "smallint")
     private int stars;
 
     @Column(length = 500)
